requirements.txt
python-dotenv
streamlit
requests

#List of req